+++ lightseq/csrc/ops_new/includes/concat3_dim1.h
@@ -11,18 +11,18 @@ class Concat3Dim1 : public Operator {
   bool _is_skip = false;
   bool _is_continuous_cache;
 
-  int _mx_sz0;
-  int _mx_sz1;
-  int _mx_sz2;
+  size_t _mx_sz0;
+  size_t _mx_sz1;
+  size_t _mx_sz2;
 
-  int _sz0;
-  int _sz1_0;
-  int _sz1_1;
+  size_t _sz0;
+  size_t _sz1_0;
+  size_t _sz1_1;
 
   Variable* _new_cache;
 
  public:
-  Concat3Dim1(int mx_sz0, int mx_sz1, int mx_sz2, bool is_continuous_cache)
+  Concat3Dim1(size_t mx_sz0, size_t mx_sz1, size_t mx_sz2, bool is_continuous_cache)
       : Operator("Concat3Dim1"),
         _mx_sz0(mx_sz0),
         _mx_sz1(mx_sz1),
@@ -33,7 +33,7 @@ class Concat3Dim1 : public Operator {
 
   Variable* operator()(Variable* inp, Variable* cache);
 
-  void before_forward(int sz0, int sz1_0, int sz1_1, bool is_skip = false) {
+  void before_forward(size_t sz0, size_t sz1_0, size_t sz1_1, bool is_skip = false) {
     _sz0 = sz0, _sz1_0 = sz1_0, _sz1_1 = sz1_1, _is_skip = is_skip;
     if (_is_continuous_cache) {
       _new_cache->set_shape({_sz0, _sz1_0 + _sz1_1, _mx_sz2});
