+++ lightseq/csrc/ops_new/includes/dropout.h
@@ -9,8 +9,8 @@ template <typename T1, typename T2>
 class DropoutOp : public Operator {
  private:
   float ratio;
-  int _max_ele_num;
-  int _count;
+  size_t _max_ele_num;
+  size_t _count;
   bool _is_skip;
 
   TensorPtr _mask;
@@ -19,7 +19,7 @@ class DropoutOp : public Operator {
  public:
   float RATIO() const { return _context_ptr->is_training() ? ratio : 0.0; }
 
-  DropoutOp(float r, int max_ele_num)
+  DropoutOp(float r, size_t max_ele_num)
       : Operator("Dropout"), ratio(r), _max_ele_num(max_ele_num) {
     _mask.reset(new Tensor("mask", g_dtype<uint8_t>(), max_ele_num));
   }
@@ -28,7 +28,7 @@ class DropoutOp : public Operator {
 
   Variable* operator()(Variable* inp);
 
-  void before_forward(int count) {
+  void before_forward(size_t count) {
     _count = count;
     _result->set_shape({count});
   }
